fix: fixed music button
