+++ client/src/components/gameboard/Gameboard.js
@@ -14,7 +14,6 @@ import failureSound from '../../assets/sound/failure.mp3';
 import successSound from '../../assets/sound/success.mp3';
 import chooseSound from '../../assets/sound/swap.mp3';
 import clickSound from '../../assets/sound/click.mp3';
-import musicSound from '../../assets/sound/3.mp3';
 
 const getRandomNumber = (min, max) => {
   const rand = min + Math.random() * (max + 1 - min);
@@ -30,12 +29,9 @@ const Gameboard = () => {
   const [playSuccess] = useSound(successSound);
   const [playChoose] = useSound(chooseSound);
   const [playClick] = useSound(clickSound);
-  const [playMusic, { stop }] = useSound(musicSound);
 
   const [soundButtonClass, setSoundButtonClass] = useState('sound-switcher red lighten-2 btn');
-  const [musicButtonClass, setMusicButtonClass] = useState('music-switcher red lighten-2 btn');
   const [isSound, setSound] = useState(false);
-  const [isMusic, setMusic] = useState(false);
   const [isChoose, setGameType] = useState(false);
   const [myChoice, setMyChoice] = useState(null);
   const [enemyThinking, setEnemyThinking] = useState(true);
@@ -52,16 +48,6 @@ const Gameboard = () => {
     document.removeEventListener('keydown', pressSoundHandler, false);
   };
 
-  const musicHandler = () => {
-    setMusic(!isMusic);
-    if (isMusic) {
-      stop();
-    } else {
-      playMusic();
-    }
-    document.removeEventListener('keydown', pressMusicHandler, false);
-  };
-
   useEffect(() => {
     if (isSound) {
       setSoundButtonClass('sound-switcher red lighten-2 btn');
@@ -71,16 +57,6 @@ const Gameboard = () => {
       message('Sound off 🙉');
     }
   }, [isSound]);
-
-  useEffect(() => {
-    if (isMusic) {
-      setMusicButtonClass('music-switcher red lighten-2 btn');
-      message('Music on 🎺');
-    } else {
-      setMusicButtonClass('music-switcher green lighten-2 btn');
-      message('Music off 🙉');
-    }
-  }, [isMusic]);
 
   const getWins = useCallback(
     async () => {
@@ -205,12 +181,6 @@ const Gameboard = () => {
     }
   }, [isSound]);
 
-  const pressMusicHandler = useCallback((e) => {
-    if (e.key === 'w') {
-      musicHandler();
-    }
-  }, [isMusic]);
-
   useEffect(() => {
     document.addEventListener('keydown', pressHandler, false);
 
@@ -234,17 +204,9 @@ const Gameboard = () => {
     };
   }, [isSound]);
 
-  useEffect(() => {
-    document.addEventListener('keydown', pressMusicHandler, false);
-    return () => {
-      document.removeEventListener('keydown', pressMusicHandler, false);
-    };
-  }, [isMusic]);
-
   if (!isChoose) {
     return (
       <div className="gameboard">
-        <button className={musicButtonClass} onClick={musicHandler} title="music On/Off"><i className="material-icons">library_music</i></button>
         <button className={soundButtonClass} onClick={soundHandler} title="sound On/Off"><i className="material-icons">surround_sound</i></button>
         <h1 className="gameboard__title white-text">Choose your weapon</h1>
         {chips.map((chip, index) => (
@@ -261,7 +223,6 @@ const Gameboard = () => {
   if (isChoose) {
     return (
       <div className="gameboard">
-        <button className={musicButtonClass} onClick={musicHandler} title="music On/Off"><i className="material-icons">library_music</i></button>
         <button className={soundButtonClass} onClick={soundHandler} title="sound On/Off"><i className="material-icons">surround_sound</i></button>
         <h1 className="gameboard__title white-text">Let`s test your luck:)</h1>
         <div className={`game-chip ${myChoice}`}
